Move utils module
